move file: lib
