+++ lib/mergeConfig.ts
